LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/border-image-slice-expected.txt


PASS Can set 'border-image-slice' to CSS-wide keywords: initial
PASS Can set 'border-image-slice' to CSS-wide keywords: inherit
PASS Can set 'border-image-slice' to CSS-wide keywords: unset
PASS Can set 'border-image-slice' to CSS-wide keywords: revert
FAIL Can set 'border-image-slice' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'border-image-slice' to a number: 0 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-slice' to a number: -3.14 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-slice' to a number: 3.14 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-slice' to a number: calc(2 + 3) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'border-image-slice' to a percent: 0% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-slice' to a percent: -3.14% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-slice' to a percent: 3.14% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-slice' to a percent: calc(0% + 0%) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
PASS Setting 'border-image-slice' to a length throws TypeError
PASS Setting 'border-image-slice' to a time throws TypeError
PASS Setting 'border-image-slice' to an angle throws TypeError
PASS Setting 'border-image-slice' to a flexible length throws TypeError
PASS Setting 'border-image-slice' to a URL throws TypeError
PASS Setting 'border-image-slice' to a transform throws TypeError
PASS 'border-image-slice' does not supported '30 fill'
PASS 'border-image-slice' does not supported '30 40 50'
PASS 'border-image-slice' does not supported '30 40 50 60 fill'
